refactor(support): extract support/netutil package from support/util
Move network visibility, route management, service publishing, networking,
and IP utility functions into a new cohesive support/netutil package.

Files moved via git mv: visibility.go, route.go, expose.go, networking.go,
public.go and their corresponding test files.
Functions extracted from util.go into new iputil.go: IsIPv4CIDR,
IsIPv4Address, FirstUsableIP, ResolveDNSHostname, HostFromURL.
Tests extracted from util_test.go into new iputil_test.go.
