+++ support/netutil/expose_test.go
@@ -1,4 +1,4 @@
-package util
+package netutil
 
 import (
 	"testing"
