+++ support/netutil/networking.go
@@ -1,4 +1,4 @@
-package util
+package netutil
 
 import (
 	"strings"
